% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{type2str}
\alias{type2str}
\title{Wrapper of Rf_type2str}
\usage{
type2str(x)
}
\arguments{
\item{x}{any SEXP object}
}
\value{
A character vector of length one
with the value returned by Rf_type2str.
}
\description{
Interact with Rf_type2str from R
}
\examples{
type2str(1L)
type2str(-12.3)
type2str("A String")

}
