% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{TYPEOF}
\alias{TYPEOF}
\title{Wrapper of TYPEOF}
\usage{
TYPEOF(x)
}
\arguments{
\item{x}{any SEXP object}
}
\value{
An integer vector of length one
encoding the numeric type value of x as an SEXP
}
\description{
Interact with TYPEOF from R
}
\examples{
TYPEOF(1L)
TYPEOF(-12.3)
TYPEOF("A String")

}
