% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{TYPEOF_list}
\alias{TYPEOF_list}
\title{Wrapper of TYPEOF with extras}
\usage{
TYPEOF_list(x)
}
\arguments{
\item{x}{any SEXP object}
}
\value{
A list of TYPEOF, Rf_type2str, and the constant symbol representing TYPEOF
}
\description{
Get information about the type of an R object
}
\examples{
TYPEOF_list(1L)
TYPEOF_list(-12.3)
TYPEOF_list("A String")

}
